Custom Metric value by creating extended querysets.
By default the count of the applied `GROUP BY` (`.values()`) and filter is used.
With this change we can define extended querysets with a CHOICES option that exposes more complex results.

This initial version has implemented the Aggregate prefix utilization in pure Django ORM.
When creating a Metric for the Aggregate model you can now additionally to "Count" select the following metric values:
- Percentage of aggregate used: this is the same value as seen on the Aggregate list/detail page.
- Total IPs in aggregate: Total IPs that are located inside the aggregate.
- IPs used in aggregate: How many IPs are occupied in the aggregate.

This change should make it easier to create metrics that are a lot more complex than a simple count.

Fixes: #5

Signed-off-by: Daan van Gorkum <djvg@djvg.net>