Prevent double quotes from custom_field_data JSON
`Cast` treats JSON values as strings and serializes them as such. This
breaks situations where the value is already a string and the result is
double quotes. Nested values are still supported.

For now this is only handled for `custom_field_data` but might be
extended if there are more JSONField use cases.

Fixes #28

Signed-off-by: Daan van Gorkum <djvg@djvg.net>