API improvements
From this commit management of the Metric(s)(Collection) is possible using
the REST API.

Before this some basic information was available through the API but
complete management was not possible.

Additionally the objects will provide a `export_url` key to link
directly to generated metrics.

Other improvements involve better validation to prevent duplicate
labels or invalid metric values to be selected.

Fixes #21

Signed-off-by: Daan van Gorkum <djvg@djvg.net>